feat: local configuration
