+++ README.md
@@ -2,22 +2,22 @@
 
 A library to show off your widget library.
 
-* Show running demos
-* Syntax highlighted example code
-* Auto documented widget properties
-* Auto documented theme class hooks
-* Run on codesandbox
-* Run unit tests
+-   Show running demos
+-   Syntax highlighted example code
+-   Auto documented widget properties
+-   Auto documented theme class hooks
+-   Run on codesandbox
+-   Run unit tests
 
-![screenshot](/screenshot.png?raw=true "screenshot")
+![screenshot](/screenshot.png?raw=true 'screenshot')
 
 ## Compatibility
 
 Each major release of Dojo Parade is compatible with a specific major version of the rest of the Dojo framework:
 
 | Parade Version | Dojo Version |
-| :------------: | :-----------:|
-| `1.*.*`        | `7.*.*`      |
+| :------------: | :----------: |
+|    `1.*.*`     |   `7.*.*`    |
 
 ## Usage
 
@@ -26,13 +26,15 @@ npm install @dojo/parade
 ```
 
 #### **`.dojorc`**
+
 ```json
 {
-  "extends": "./node_modules/@dojo/parade/parade.json"
+	"extends": "./node_modules/@dojo/parade/parade.json"
 }
 ```
 
 #### **`main.tsx`**
+
 ```tsx
 import parade from '@dojo/parade';
 
@@ -42,14 +44,15 @@ import config from './config';
 parade({ config });
 ```
 
-### Example config:
+### Example global config:
+
 #### **`config.tsx`**
+
 ```tsx
 import dojoTheme from '@dojo/themes/dojo';
 import '@dojo/themes/dojo/index.css';
 
 import BasicAccordion from './widgets/accordion/Basic';
-import Exclusive from './widgets/accordion/Exclusive';
 import BasicButton from './widgets/button/Basic';
 import DisabledSubmit from './widgets/button/DisabledSubmit';
 import ToggleButton from './widgets/button/ToggleButton';
@@ -62,21 +65,6 @@ export default {
   tests,
   themes: [ dojoTheme ],
   widgets: {
-    accordion: {
-      examples: [
-        {
-          filename: 'Exclusive',
-          module: Exclusive
-        }
-      ],
-      filename: 'index',
-      overview: {
-        example: {
-          filename: 'Basic',
-          module: BasicAccordion
-        }
-      }
-    },
     button: {
       examples: [
         {
@@ -99,4 +87,32 @@ export default {
       }
     }
   }
+```
+
+### Example local config:
+
+#### **`accordion.example.tsx`**
+
+```tsx
+import dojoTheme from '@dojo/themes/dojo';
+import '@dojo/themes/dojo/index.css';
+
+import BasicAccordion from './widgets/accordion/Basic';
+import Exclusive from './widgets/accordion/Exclusive';
+
+export default {
+	examples: [
+		{
+			filename: 'Exclusive',
+			module: Exclusive
+		}
+	],
+	filename: 'index',
+	overview: {
+		example: {
+			filename: 'Basic',
+			module: BasicAccordion
+		}
+	}
+};
 ```
