+++ src/example/DefaultClassButton/DefaultClassButton.example.ts
@@ -1,0 +1,11 @@
+import DefaultClassButtonExample from './examples/DefaultClassButtonExample';
+
+export default {
+	filename: 'DefaultClassButton',
+	overview: {
+		example: {
+			filename: 'DefaultClassButtonExample',
+			module: DefaultClassButtonExample
+		}
+	}
+};
