+++ src/example/InheritingButton/InheritingButton.example.ts
@@ -1,0 +1,11 @@
+import InheritingButtonExample from './examples/InheritingButtonExample';
+
+export default {
+	filename: 'InheritingButton',
+	overview: {
+		example: {
+			filename: 'InheritingButtonExample',
+			module: InheritingButtonExample
+		}
+	}
+};
