+++ src/example/NamedButton/NamedButton.example.ts
@@ -1,0 +1,11 @@
+import NamedButtonExample from './examples/NamedButtonExample';
+
+export default {
+	filename: 'NamedButton',
+	overview: {
+		example: {
+			filename: 'NamedButtonExample',
+			module: NamedButtonExample
+		}
+	}
+};
