+++ src/example/NamedButton/examples/NamedButtonExample.tsx
@@ -1,8 +1,8 @@
 import { create, tsx } from '@dojo/framework/core/vdom';
-import Button from './NamedClassButton/NamedClassButton';
+import Button from '../NamedButton';
 
 const factory = create();
 
 export default factory(function Basic() {
-	return <Button label="Hello" />;
+	return <Button>Hello</Button>;
 });
