+++ src/example/NamedClassButton/NamedClassButton.example.ts
@@ -1,0 +1,17 @@
+import NamedClassButtonExample from './examples/NamedClassButtonExample';
+import { create } from '@dojo/framework/core/vdom';
+
+const ExampleDescription = create()(
+	() => 'This is an example of an optional description that is a widget.'
+);
+
+export default {
+	filename: 'NamedClassButton',
+	overview: {
+		example: {
+			filename: 'NamedClassButtonExample',
+			module: NamedClassButtonExample,
+			description: ExampleDescription
+		}
+	}
+};
