+++ src/example/NamedClassButton/examples/NamedClassButtonExample.tsx
@@ -1,8 +1,8 @@
 import { create, tsx } from '@dojo/framework/core/vdom';
-import Button from './input/Input';
+import Button from '../NamedClassButton';
 
 const factory = create();
 
 export default factory(function Basic() {
-	return <Button label="Hello" value={''} onValue={() => {}} type="labeled" />;
+	return <Button label="Hello" />;
 });
