+++ src/example/input/Input.example.ts
@@ -1,0 +1,11 @@
+import InputExample from './examples/InputExample';
+
+export default {
+	filename: 'Input',
+	overview: {
+		example: {
+			filename: 'InputExample',
+			module: InputExample
+		}
+	}
+};
